LPS-28160 Support for recycle bin in journal folders

+++ portal-impl/src/com/liferay/portlet/journal/service/impl/JournalFolderLocalServiceImpl.java
@@ -23,6 +23,7 @@ import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.IndexerRegistryUtil;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.OrderByComparator;
+import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.model.ResourceConstants;
 import com.liferay.portal.model.User;
@@ -37,6 +38,8 @@ import com.liferay.portlet.journal.model.JournalArticle;
 import com.liferay.portlet.journal.model.JournalFolder;
 import com.liferay.portlet.journal.model.JournalFolderConstants;
 import com.liferay.portlet.journal.service.base.JournalFolderLocalServiceBaseImpl;
+import com.liferay.portlet.trash.model.TrashEntry;
+import com.liferay.portlet.trash.util.TrashUtil;
 
 import java.util.ArrayList;
 import java.util.Date;
@@ -97,13 +100,23 @@ public class JournalFolderLocalServiceImpl
 	public JournalFolder deleteFolder(JournalFolder folder)
 		throws PortalException, SystemException {
 
+		return deleteFolder(folder, true);
+	}
+
+	@Indexable(type = IndexableType.DELETE)
+	public JournalFolder deleteFolder(
+			JournalFolder folder, boolean includeTrashedEntries)
+		throws PortalException, SystemException {
+
 		// Folders
 
 		List<JournalFolder> folders = journalFolderPersistence.findByG_P(
 			folder.getGroupId(), folder.getFolderId());
 
 		for (JournalFolder curFolder : folders) {
-			deleteFolder(curFolder);
+			if (includeTrashedEntries || !curFolder.isInTrash()) {
+				deleteFolder(curFolder);
+			}
 		}
 
 		// Folder
@@ -118,11 +131,16 @@ public class JournalFolderLocalServiceImpl
 		// Entries
 
 		journalArticleLocalService.deleteArticles(
-			folder.getGroupId(), folder.getFolderId());
+			folder.getGroupId(), folder.getFolderId(), includeTrashedEntries);
 
 		// Expando
 
 		expandoValueLocalService.deleteValues(
+			JournalFolder.class.getName(), folder.getFolderId());
+
+		// Trash
+
+		trashEntryLocalService.deleteEntry(
 			JournalFolder.class.getName(), folder.getFolderId());
 
 		return folder;
@@ -135,7 +153,18 @@ public class JournalFolderLocalServiceImpl
 		JournalFolder folder = journalFolderPersistence.findByPrimaryKey(
 			folderId);
 
-		return deleteFolder(folder);
+		return deleteFolder(folder, true);
+	}
+
+	@Indexable(type = IndexableType.DELETE)
+	public JournalFolder deleteFolder(
+			long folderId, boolean includeTrashedEntries)
+		throws PortalException, SystemException {
+
+		JournalFolder folder = journalFolderPersistence.findByPrimaryKey(
+			folderId);
+
+		return deleteFolder(folder, includeTrashedEntries);
 	}
 
 	public void deleteFolders(long groupId)
@@ -147,6 +176,14 @@ public class JournalFolderLocalServiceImpl
 		for (JournalFolder folder : folders) {
 			journalFolderLocalService.deleteFolder(folder);
 		}
+	}
+
+	public JournalFolder fetchFolder(
+			long groupId, long parentFolderId, String name)
+		throws SystemException {
+
+		return journalFolderPersistence.fetchByG_P_N(
+				groupId, parentFolderId, name);
 	}
 
 	public JournalFolder fetchFolder(long groupId, String name)
@@ -179,15 +216,43 @@ public class JournalFolderLocalServiceImpl
 	public List<JournalFolder> getFolders(long groupId, long parentFolderId)
 		throws SystemException {
 
-		return journalFolderPersistence.findByG_P(groupId, parentFolderId);
+		return getFolders(
+			groupId, parentFolderId, WorkflowConstants.STATUS_APPROVED);
+	}
+
+	public List<JournalFolder> getFolders(
+			long groupId, long parentFolderId, int status)
+		throws SystemException {
+
+		return journalFolderPersistence.findByG_P_S(
+			groupId, parentFolderId, status);
 	}
 
 	public List<JournalFolder> getFolders(
 			long groupId, long parentFolderId, int start, int end)
 		throws SystemException {
 
-		return journalFolderPersistence.findByG_P(
-			groupId, parentFolderId, start, end);
+		return getFolders(
+			groupId, parentFolderId, WorkflowConstants.STATUS_APPROVED, start,
+			end);
+	}
+
+	public List<JournalFolder> getFolders(
+			long groupId, long parentFolderId, int status, int start, int end)
+		throws SystemException {
+
+		return journalFolderPersistence.findByG_P_S(
+				groupId, parentFolderId, status, start, end);
+	}
+
+	public List<Object> getFoldersAndArticles(long groupId, long folderId)
+		throws SystemException {
+
+		QueryDefinition queryDefinition = new QueryDefinition(
+			WorkflowConstants.STATUS_ANY);
+
+		return journalFolderFinder.findF_A_ByG_F(
+			groupId, folderId, queryDefinition);
 	}
 
 	public List<Object> getFoldersAndArticles(
@@ -239,7 +304,15 @@ public class JournalFolderLocalServiceImpl
 	public int getFoldersCount(long groupId, long parentFolderId)
 		throws SystemException {
 
-		return journalFolderPersistence.countByG_P(groupId, parentFolderId);
+		return getFoldersCount(
+			groupId, parentFolderId, WorkflowConstants.STATUS_APPROVED);
+	}
+
+	public int getFoldersCount(long groupId, long parentFolderId, int status)
+		throws SystemException {
+
+		return journalFolderPersistence.countByG_P_S(
+			groupId, parentFolderId, status);
 	}
 
 	public List<JournalFolder> getNoAssetFolders() throws SystemException {
@@ -282,6 +355,59 @@ public class JournalFolderLocalServiceImpl
 		journalFolderPersistence.update(folder);
 
 		return folder;
+	}
+
+	public JournalFolder moveFolderFromTrash(
+			long userId, long folderId, long parentFolderId,
+			ServiceContext serviceContext)
+		throws PortalException, SystemException {
+
+		JournalFolder folder = journalFolderPersistence.findByPrimaryKey(
+			folderId);
+
+		if (folder.isInTrash()) {
+			restoreFolderFromTrash(userId, folderId);
+		}
+		else {
+			updateStatus(userId, folder, WorkflowConstants.STATUS_APPROVED);
+		}
+
+		return moveFolder(folderId, parentFolderId, serviceContext);
+	}
+
+	public void moveFolderToTrash(long userId, long folderId)
+		throws PortalException, SystemException {
+
+		JournalFolder folder = journalFolderPersistence.findByPrimaryKey(
+			folderId);
+
+		folder = updateStatus(
+			userId, folder, WorkflowConstants.STATUS_IN_TRASH);
+
+		TrashEntry trashEntry = trashEntryLocalService.getEntry(
+			JournalFolder.class.getName(), folder.getFolderId());
+
+		String trashTitle = TrashUtil.getTrashTitle(trashEntry.getEntryId());
+
+		folder.setName(trashTitle);
+
+		journalFolderPersistence.update(folder);
+	}
+
+	public void restoreFolderFromTrash(long userId, long folderId)
+		throws PortalException, SystemException {
+
+		JournalFolder folder = journalFolderPersistence.findByPrimaryKey(
+			folderId);
+
+		folder.setName(TrashUtil.getOriginalTitle(folder.getName()));
+
+		journalFolderPersistence.update(folder);
+
+		TrashEntry trashEntry = trashEntryLocalService.getEntry(
+			JournalFolder.class.getName(), folderId);
+
+		updateStatus(userId, folder, trashEntry.getStatus());
 	}
 
 	public void updateAsset(
@@ -340,6 +466,55 @@ public class JournalFolderLocalServiceImpl
 			userId, folder, serviceContext.getAssetCategoryIds(),
 			serviceContext.getAssetTagNames(),
 			serviceContext.getAssetLinkEntryIds());
+
+		return folder;
+	}
+
+	public JournalFolder updateStatus(
+			long userId, JournalFolder folder, int status)
+		throws PortalException, SystemException {
+
+		User user = userPersistence.findByPrimaryKey(userId);
+
+		int oldStatus = folder.getStatus();
+
+		folder.setStatus(status);
+		folder.setStatusByUserId(userId);
+		folder.setStatusByUserName(user.getFullName());
+		folder.setStatusDate(new Date());
+
+		journalFolderPersistence.update(folder);
+
+		// Folders and entries
+
+		List<Object> foldersAndEntries =
+			journalFolderLocalService.getFoldersAndArticles(
+				folder.getGroupId(), folder.getFolderId());
+
+		updateDependentStatus(foldersAndEntries, status);
+
+		// Trash
+
+		if (oldStatus == WorkflowConstants.STATUS_IN_TRASH) {
+			trashEntryLocalService.deleteEntry(
+				JournalFolder.class.getName(), folder.getFolderId());
+		}
+		else if (status == WorkflowConstants.STATUS_IN_TRASH) {
+			UnicodeProperties typeSettingsProperties = new UnicodeProperties();
+
+			typeSettingsProperties.put("title", folder.getName());
+
+			trashEntryLocalService.addTrashEntry(
+				userId, folder.getGroupId(), JournalFolder.class.getName(),
+				folder.getFolderId(), oldStatus, null, typeSettingsProperties);
+		}
+
+		// Index
+
+		Indexer indexer = IndexerRegistryUtil.nullSafeGetIndexer(
+			JournalFolder.class);
+
+		indexer.reindex(folder);
 
 		return folder;
 	}
@@ -420,6 +595,107 @@ public class JournalFolderLocalServiceImpl
 		}
 
 		journalFolderLocalService.deleteFolder(fromFolder);
+	}
+
+	protected void updateDependentStatus(
+			List<Object> foldersAndEntries, int status)
+		throws PortalException, SystemException {
+
+		for (Object object : foldersAndEntries) {
+			if (object instanceof JournalArticle) {
+				JournalArticle article = (JournalArticle)object;
+
+				if (status == WorkflowConstants.STATUS_IN_TRASH) {
+
+					// Asset
+
+					if (article.getStatus() ==
+							WorkflowConstants.STATUS_APPROVED) {
+
+						assetEntryLocalService.updateVisible(
+							JournalArticle.class.getName(),
+							article.getResourcePrimKey(), false);
+					}
+
+					// Social
+
+					socialActivityCounterLocalService.disableActivityCounters(
+						JournalArticle.class.getName(),
+						article.getResourcePrimKey());
+
+					if (article.getStatus() ==
+							WorkflowConstants.STATUS_PENDING) {
+
+						article.setStatus(WorkflowConstants.STATUS_DRAFT);
+
+						journalArticlePersistence.update(article);
+					}
+				}
+				else {
+
+					// Asset
+
+					if (article.getStatus() ==
+							WorkflowConstants.STATUS_APPROVED) {
+
+						assetEntryLocalService.updateVisible(
+							JournalArticle.class.getName(),
+							article.getResourcePrimKey(), true);
+					}
+
+					// Social
+
+					socialActivityCounterLocalService.enableActivityCounters(
+						JournalArticle.class.getName(),
+						article.getResourcePrimKey());
+				}
+
+				// Workflow
+
+				if (status != WorkflowConstants.STATUS_APPROVED) {
+					List<JournalArticle> articleVersions =
+						journalArticlePersistence.findByG_A(
+							article.getGroupId(), article.getArticleId());
+
+					for (JournalArticle curArticle : articleVersions) {
+						if (!curArticle.isPending()) {
+							continue;
+						}
+
+						curArticle.setStatus(WorkflowConstants.STATUS_DRAFT);
+
+						journalArticlePersistence.update(curArticle);
+
+						workflowInstanceLinkLocalService.
+							deleteWorkflowInstanceLink(
+								curArticle.getCompanyId(),
+								curArticle.getGroupId(),
+								JournalArticle.class.getName(),
+								curArticle.getId());
+					}
+				}
+
+				// Indexer
+
+				Indexer indexer = IndexerRegistryUtil.nullSafeGetIndexer(
+					JournalArticle.class);
+
+				indexer.reindex(article);
+			}
+			else if (object instanceof JournalFolder) {
+				JournalFolder folder = (JournalFolder)object;
+
+				if (folder.isInTrash()) {
+					continue;
+				}
+
+				List<Object> curFoldersAndEntries =
+					journalFolderLocalService.getFoldersAndArticles(
+						folder.getGroupId(), folder.getFolderId());
+
+				updateDependentStatus(curFoldersAndEntries, status);
+			}
+		}
 	}
 
 	protected void validateFolder(
